# frontend/packages/api/openapi/admin.json
{
  "swagger": "2.0",
  "info": {
    "description": "基于Gin + Vue + Element UI的前后端分离权限管理系统的接口文档\n添加qq群: 521386980 进入技术交流群 请先star，谢谢！",
    "title": "go-admin API",
    "contact": {},
    "license": {
      "name": "MIT",
      "url": "https://github.com/go-admin-team/go-admin/blob/master/LICENSE.md"
    },
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/dict/type": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "获取JSON",
        "tags": [
          "字典类型"
        ],
        "summary": "字典类型列表数据",
        "parameters": [
          {
            "type": "string",
            "description": "dictName",
            "name": "dictName",
            "in": "query"
          },
          {
            "type": "string",
            "description": "dictId",
            "name": "dictId",
            "in": "query"
          },
          {
            "type": "string",
            "description": "dictType",
            "name": "dictType",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "页条数",
            "name": "pageSize",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "页码",
            "name": "pageIndex",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": 200, \"data\": [...]}",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "获取JSON",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "字典类型"
        ],
        "summary": "添加字典类型",
        "parameters": [
          {
            "description": "data",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.SysDictTypeInsertReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": 200, \"data\": [...]}",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "删除数据",
        "tags": [
          "字典类型"
        ],
        "summary": "删除字典类型",
        "parameters": [
          {
            "description": "body",
            "name": "dictCode",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.SysDictTypeDeleteReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": 200, \"data\": [...]}",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/api/v1/dict/type-option-select": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "获取JSON",
        "tags": [
          "字典类型"
        ],
        "summary": "字典类型全部数据",
        "parameters": [
          {
            "type": "string",
            "description": "dictName",
            "name": "dictName",
            "in": "query"
          },
          {
            "type": "string",
            "description": "dictId",
            "name": "dictId",
            "in": "query"
          },
          {
            "type": "string",
            "description": "dictType",
            "name": "dictType",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": 200, \"data\": [...]}",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/api/v1/dict/type/{dictId}": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "获取JSON",
        "tags": [
          "字典类型"
        ],
        "summary": "字典类型通过字典id获取",
        "parameters": [
          {
            "type": "integer",
            "description": "字典类型编码",
            "name": "dictId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": 200, \"data\": [...]}",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      },
      "put": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "获取JSON",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "字典类型"
        ],
        "summary": "修改字典类型",
        "parameters": [
          {
            "type": "integer",
            "description": "字典类型编码",
            "name": "dictId",
            "in": "path",
            "required": true
          },
          {
            "description": "body",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.SysDictTypeUpdateReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": 200, \"data\": [...]}",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
...
    "/api/v1/sys-api": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "获取接口管理列表",
        "tags": [
          "接口管理"
        ],
        "summary": "获取接口管理列表",
        "parameters": [
          {
            "type": "string",
            "description": "名称",
            "name": "name",
            "in": "query"
          },
          {
            "type": "string",
            "description": "标题",
            "name": "title",
            "in": "query"
          },
          {
            "type": "string",
            "description": "地址",
            "name": "path",
            "in": "query"
          },
          {
            "type": "string",
            "description": "类型",
            "name": "action",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "页条数",
            "name": "pageSize",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "页码",
            "name": "pageIndex",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": 200, \"data\": [...]}",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "allOf": [
                        {
                          "$ref": "#/definitions/response.Page"
                        },
                        {
                          "type": "object",
                          "properties": {
                            "list": {
                              "type": "array",
                              "items": {
                                "$ref": "#/definitions/go-admin_app_admin_models.SysApi"
                              }
                            }
                          }
                        }
                      ]
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "删除接口管理",
        "tags": [
          "接口管理"
        ],
        "summary": "删除接口管理",
        "parameters": [
          {
            "description": "body",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.SysApiDeleteReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": 200, \"message\": \"删除成功\"}",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
    "/api/v1/sys-api/{id}": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "获取接口管理",
        "tags": [
          "接口管理"
        ],
        "summary": "获取接口管理",
        "parameters": [
          {
            "type": "string",
            "description": "id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": 200, \"data\": [...]}",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/go-admin_app_admin_models.SysApi"
                    }
                  }
                }
              ]
            }
          }
        }
      },
      "put": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "修改接口管理",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "接口管理"
        ],
        "summary": "修改接口管理",
        "parameters": [
          {
            "type": "string",
            "description": "id",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "body",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.SysApiUpdateReq"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{\"code\": 200, \"message\": \"修改成功\"}",
            "schema": {
              "$ref": "#/definitions/response.Response"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "dto.SysApiDeleteReq": {
      "type": "object",
      "properties": {
        "ids": {
          "type": "array",
          "items": {
            "type": "integer"
          }
        }
      }
    },
    "dto.SysApiUpdateReq": {
      "type": "object",
      "properties": {
        "action": {
          "type": "string"
        },
        "createBy": {
          "type": "integer"
        },
        "handle": {
          "type": "string"
        },
        "id": {
          "description": "编码",
          "type": "integer"
        },
        "path": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "updateBy": {
          "type": "integer"
        }
      }
    },
...
    "dto.SysDictTypeDeleteReq": {
      "type": "object",
      "properties": {
        "createBy": {
          "type": "integer"
        },
        "ids": {
          "type": "array",
          "items": {
            "type": "integer"
          }
        },
        "updateBy": {
          "type": "integer"
        }
      }
    },
    "dto.SysDictTypeInsertReq": {
      "type": "object",
      "properties": {
        "createBy": {
          "type": "integer"
        },
        "dictName": {
          "type": "string"
        },
        "dictType": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "remark": {
          "type": "string"
        },
        "status": {
          "type": "integer"
        },
        "updateBy": {
          "type": "integer"
        }
      }
    },
    "dto.SysDictTypeUpdateReq": {
      "type": "object",
      "properties": {
        "createBy": {
          "type": "integer"
        },
        "dictName": {
          "type": "string"
        },
        "dictType": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "remark": {
          "type": "string"
        },
        "status": {
          "type": "integer"
        },
        "updateBy": {
          "type": "integer"
        }
      }
    },
...
    "go-admin_app_admin_models.SysApi": {
      "type": "object",
      "properties": {
        "action": {
          "type": "string"
        },
        "createBy": {
          "type": "integer"
        },
        "createdAt": {
          "type": "string"
        },
        "handle": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "path": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "updateBy": {
          "type": "integer"
        },
        "updatedAt": {
          "type": "string"
        }
      }
    },
...
    "response.Page": {
      "type": "object",
      "properties": {
        "count": {
          "type": "integer"
        },
        "pageIndex": {
          "type": "integer"
        },
        "pageSize": {
          "type": "integer"
        }
      }
    },
    "response.Response": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer"
        },
        "msg": {
          "type": "string"
        },
        "requestId": {
          "description": "数据集",
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}